chore(admin): convert Single Sign On settings page to TS and refactoring
