+++ examples/getstarted/src/api/address/content-types/address/schema.json
@@ -76,3 +76,4 @@
     }
   }
 }
+
