+++ packages/core/admin/admin/src/types/permissions.ts
@@ -1,6 +1,6 @@
 import type { Permission } from '@strapi/helper-plugin';
 
-type SettingsPermissions = 'project-settings' | 'roles' | 'users' | 'webhooks';
+type SettingsPermissions = 'project-settings' | 'users' | 'webhooks' | 'sso' | 'roles';
 
 interface CRUDPermissions {
   main: Permission[];
